+++ bibfor/algeline/chcoma.F90
@@ -63,7 +63,7 @@ subroutine chcoma(tablez, nomaou)
 !                                 ROUTINE.
 !.========================= DEBUT DES DECLARATIONS ====================
 ! -----  VARIABLES LOCALES
-    integer :: ngm, ibid, iret, idcode, dimcoo, nbno, jcoor, idcoor, ino
+    integer(kind=8) :: ngm, ibid, iret, idcode, dimcoo, nbno, jcoor, idcoor, ino
     real(kind=8) :: r8b, p(2, 2), alpha, yg, zg, yabs, zabs
     complex(kind=8) :: c16b
     character(len=8) :: k8b, typobj
