+++ bibfor/algeline/colni2.F90
@@ -20,10 +20,10 @@ subroutine colni2(col1, col2, n, d1, d2, &
                   coef1, t1, t2, eps, ier)
 ! person_in_charge: olivier.boiteau at edf.fr
     implicit none
-    integer :: n, ier
+    integer(kind=8) :: n, ier
     real(kind=8) :: col1(n), col2(n), d1, d2, coef1, t1(n), t2(n), eps
 !
-    integer :: i
+    integer(kind=8) :: i
 !-----------------------------------------------------------------------
 !-----------------------------------------------------------------------
     if (abs(d2) .le. eps) then
