+++ bibfor/algeline/crsvld.F90
@@ -27,7 +27,7 @@ subroutine crsvld(motfac, solveu, istop, nprec, &
 #include "asterfort/jeveuo.h"
 #include "asterfort/jevtbl.h"
 !
-    integer :: istop, nprec
+    integer(kind=8) :: istop, nprec
     real(kind=8) :: epsmat
     character(len=3) :: mixpre, kellag
     character(len=8) :: kxfem
@@ -48,9 +48,9 @@ subroutine crsvld(motfac, solveu, istop, nprec, &
 !
 !
 !
-    integer :: ibid
+    integer(kind=8) :: ibid
     character(len=8) :: renum
-    integer, pointer :: slvi(:) => null()
+    integer(kind=8), pointer :: slvi(:) => null()
     real(kind=8), pointer :: slvr(:) => null()
     character(len=24), pointer :: slvk(:) => null()
 !
